Created banner and configured script to accept words longer than 6 chars.
Banner created for the script.
Added code which breakdown the user input into groups of 6 characters.
Each group of characters are ran through the script and then combined at
the end of the script. For example, the word "stevens" will be split into
two strings, "steven" and "s". Each string will be leeted and then
each permutation will be combined to create strings of the original input
length which was 7.